+++ frontend/src/views/Footer.vue
@@ -6,9 +6,9 @@ const { t } = useScopedI18n('views.Footer')
 </script>
 
 <template>
-    <div>
+    <div class="app-footer">
         <n-divider class="footer-divider" />
-        <div style="text-align: center; padding: 20px">
+        <div class="footer-content">
             <n-space justify="center">
                 <n-text depth="3">
                     {{ t('copyright') }} © 2023-2026 GetAnEmail
@@ -20,8 +20,17 @@ const { t } = useScopedI18n('views.Footer')
 
 
 <style scoped>
+.app-footer {
+    padding-bottom: 10px;
+}
+
 .footer-divider {
     margin: 0;
     padding: 0 var(--x-padding);
+}
+
+.footer-content {
+    padding: 18px 20px 22px;
+    text-align: center;
 }
 </style>